pyproject.toml
[tool.poetry]
name = "tiny-index"
version = "0.1.0"
description = ""
authors = ["Takahiko Ito <takahiko.ito@dr-ubie.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Janome = "^0.4.2"
wanakana-python = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
